refactor: migration types
